almost there

+++ hello_world/6-size.c
@@ -8,7 +8,7 @@ int main(void)
 printf("Size of a char: %ld", sizeof(char));
 printf("\nSize of an int: %ld", sizeof(int));
 printf("\nSize of a long int: %ld", sizeof(long int));
-printf("\nSize of a long long int: %ld", sizeof(long double));
+printf("\nSize of a long long int: %ld", sizeof(long long int));
 printf("\nSize of a float: %ld", sizeof(float));
 return (0);
 }
